+++ Ui/Admin/Action/PortalEntityListUi.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Ui\Admin\Action;
 
+use Heptacom\HeptaConnect\Core\Ui\Admin\Audit\Contract\AuditTrailFactoryInterface;
 use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult;
@@ -13,13 +14,19 @@ use Heptacom\HeptaConnect\Ui\Admin\Base\Action\UiActionType;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\Portal\PortalEntityListUiActionInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\PortalNode\PortalNodeEntityListUiActionInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionContextInterface;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\ReadException;
 
 final class PortalEntityListUi implements PortalEntityListUiActionInterface
 {
+    private AuditTrailFactoryInterface $auditTrailFactory;
+
     private PortalNodeEntityListUiActionInterface $portalNodeEntityListUiAction;
 
-    public function __construct(PortalNodeEntityListUiActionInterface $portalNodeEntityListUiAction)
-    {
+    public function __construct(
+        AuditTrailFactoryInterface $auditTrailFactory,
+        PortalNodeEntityListUiActionInterface $portalNodeEntityListUiAction
+    ) {
+        $this->auditTrailFactory = $auditTrailFactory;
         $this->portalNodeEntityListUiAction = $portalNodeEntityListUiAction;
     }
 
@@ -30,6 +37,7 @@ final class PortalEntityListUi implements PortalEntityListUiActionInterface
 
     public function list(PortalEntityListCriteria $criteria, UiActionContextInterface $context): iterable
     {
+        $trail = $this->auditTrailFactory->create($this, $context->getAuditContext(), [$criteria, $context]);
         $portalNodeKey = new PreviewPortalNodeKey($criteria->getPortal());
 
         $portalNodeCriteria = new PortalNodeEntityListCriteria($portalNodeKey);
@@ -38,13 +46,17 @@ final class PortalEntityListUi implements PortalEntityListUiActionInterface
         $portalNodeCriteria->setShowReceiver($criteria->getShowReceiver());
         $portalNodeCriteria->setFilterSupportedEntityType($criteria->getFilterSupportedEntityType());
 
-        return \iterable_map(
-            $this->portalNodeEntityListUiAction->list($portalNodeCriteria, $context),
-            static fn (PortalNodeEntityListResult $result) => new PortalEntityListResult(
-                $result->getCodeOrigin(),
-                $result->getSupportedEntityType(),
-                $result->getFlowComponentClass()
-            )
-        );
+        try {
+            yield from $trail->returnIterable(\iterable_map(
+                $this->portalNodeEntityListUiAction->list($portalNodeCriteria, $context),
+                static fn (PortalNodeEntityListResult $result) => new PortalEntityListResult(
+                    $result->getCodeOrigin(),
+                    $result->getSupportedEntityType(),
+                    $result->getFlowComponentClass()
+                )
+            ));
+        } catch (\Throwable $throwable) {
+            throw $trail->throwable(new ReadException(1663051795, $throwable));
+        }
     }
 }