+++ Ui/Admin/Action/RouteAddUiDefault.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Ui\Admin\Action;
 
+use Heptacom\HeptaConnect\Core\Ui\Admin\Audit\Contract\AuditTrailFactoryInterface;
 use Heptacom\HeptaConnect\Storage\Base\Enum\RouteCapability;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Route\RouteAdd\RouteAddDefault;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\UiActionType;
@@ -12,6 +13,13 @@ use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionContextInterface
 
 final class RouteAddUiDefault implements RouteAddUiDefaultProviderInterface
 {
+    private AuditTrailFactoryInterface $auditTrailFactory;
+
+    public function __construct(AuditTrailFactoryInterface $auditTrailFactory)
+    {
+        $this->auditTrailFactory = $auditTrailFactory;
+    }
+
     public static function class(): UiActionType
     {
         return new UiActionType(RouteAddUiDefaultProviderInterface::class);
@@ -19,6 +27,8 @@ final class RouteAddUiDefault implements RouteAddUiDefaultProviderInterface
 
     public function getDefault(UiActionContextInterface $context): RouteAddDefault
     {
-        return new RouteAddDefault(RouteCapability::ALL);
+        $trail = $this->auditTrailFactory->create($this, $context->getAuditContext(), [$context]);
+
+        return $trail->return(new RouteAddDefault(RouteCapability::ALL));
     }
 }