+++ Ui/Admin/Audit/AuditTrail.php
@@ -17,6 +17,24 @@ final class AuditTrail implements AuditTrailInterface
         }
     }
 
+    public function return(object $result): object
+    {
+        $this->end();
+
+        return $result;
+    }
+
+    public function returnIterable(iterable $result): iterable
+    {
+        try {
+            yield from $result;
+
+            $this->end();
+        } catch (\Throwable $throwable) {
+            throw $this->throwable($throwable);
+        }
+    }
+
     public function throwable(\Throwable $throwable): \Throwable
     {
         $this->end();